Auto-Coder: Address issue #460
